LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/module/instantiation-error-2-expected.txt

CONSOLE MESSAGE: SyntaxError: Importing binding name 'default' cannot be resolved by star export entries.
CONSOLE MESSAGE: SyntaxError: Importing binding name 'default' cannot be resolved by star export entries.
CONSOLE MESSAGE: SyntaxError: Importing binding name 'default' cannot be resolved by star export entries.
CONSOLE MESSAGE: SyntaxError: Importing binding name 'default' cannot be resolved by star export entries.
CONSOLE MESSAGE: SyntaxError: Importing binding name 'default' cannot be resolved by star export entries.

PASS Test that missing exports lead to SyntaxError events on window and load events on script
